requirements.txt
eel
wolframalpha
pyttsx3
SpeechRecognition
PyAudio
pyautogui
pywhatkit
pvporcupine
playsound==1.2.2
hugchat
requests
beautifulsoup4
speedtest-cli
opencv-python
mediapipe
numpy
python-dotenv
mss
python-docx
python-pptx
google-generativeai
